+++ components/projects/ProjectBox.tsx
@@ -2,8 +2,9 @@ import { FunctionComponent, useState, useEffect } from 'react';
 import { Grid, makeStyles, createStyles } from '@material-ui/core';
 import { Student } from '../../interfaces';
 import Link from 'next/link';
-import { gql } from 'apollo-boost';
+import { GET_REPO_INFO } from '../../gql/queries/projects';
 import { useQuery } from '@apollo/react-hooks';
+import { getUserRepoName, formatDate, projectStanding, dataExtract } from './helpers';
 
 type ProjectBoxProps = {
     name: string,
@@ -13,23 +14,7 @@ type ProjectBoxProps = {
     deleteMode: boolean
 }
 
-const GET_REPO_INFO = gql`
-query getRepoInfo($ownerName: String!, $repoName: String!)  {
-    repository(owner: $ownerName, name: $repoName) {
-      object(expression: "master") {
-        ... on Commit {
-          history(first: 1) {
-            totalCount
-            nodes {
-              message
-              authoredDate
-            }
-          }
-        }
-      }
-    }
-  }  
-`;
+
 const useStyles = makeStyles(() => createStyles({
    box: {
        textAlign: 'center',
@@ -80,7 +65,8 @@ const ProjectBox: FunctionComponent<ProjectBoxProps> = ({name, projectId, github
     }, [data])
     if(data)
         return(
-            <Grid key = {projectId} container item className = {deleteMode ? [classes.box, classes.deleteHover].join(' ') :[classes.box, projectStanding(lastCommitDate, classes)].join(' ')  } justify = 'center' alignItems = 'center'>
+            <Grid key = {projectId} container item className = {deleteMode ? [classes.box, classes.deleteHover].join(' ') :[classes.box, projectStanding(lastCommitDate, classes)].join(' ')} 
+            justify = 'center' alignItems = 'center'>
                 <Link href = {`/projects/${projectId}`}>
                     <Grid item xs = {3} >
                         {name}
@@ -105,32 +91,3 @@ const ProjectBox: FunctionComponent<ProjectBoxProps> = ({name, projectId, github
         return(<p>Loading</p>)
 }
 export default ProjectBox;
-
-function getUserRepoName(url: string) {
-    const userRepo = url.slice(19);
-    return userRepo.split('/');
-}
-
-function dataExtract(data: any) {
-    const result = {
-        num: data.repository.object.history.totalCount,
-        date: data.repository.object.history.nodes[0].authoredDate
-    }
-    return result;
-}
- function formatDate(commitDate: string) {
-    const date = new Date(commitDate);
-    return `${date.getDate()}.${date.getMonth()+1}.${date.getFullYear()}`
- }
-
-function projectStanding(date: string, classes: any) {
-    const commitDate = new Date(date).getTime();
-    const currentDate = new Date().getTime();
-    const elapsedTime = currentDate - commitDate;
-    if(elapsedTime<6.048e+8)
-        return classes.good;
-    else if(elapsedTime> 6.048e+8 && elapsedTime <1.814e+9)
-        return classes.warning;
-    else
-        return classes.bad;
-}
